Rework client/server payload passing code
Clean up how FileDbs are loaded/created, and add some passthrough methods to OnDiskPersistance. Server and Client now have basic semi-working implementations, but most of the work here was just figuring out how to structure the request/response cycle between a client and a server. Databases still load in a blocking manner; this needs to be fixed.